+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added support for HDR cubemaps when using `ImageBasedLighting` or `Skybox` (encoded using the RGBE8 format).
+- Added support for changing the exposure of a `Skybox`.
+- `Skybox` is now rendered with gamma correction.
+- Added request options to `glTFAsset.fromURL`.
+
 ## [2.0.1] - 2022-11-12
 ### Fixed
 - Fixed an issue which caused a crash when rendering shadows with skinned meshes.